Rename file from default to errorManifest
